Fix attachment_url vs url problem
In my Wordpress file, the GUID did not match the URL for the files... this uncovered a bug in the importer.
The image URL is not found in "attachment_url", but in the field "url". I'm not sure if this is a break between different WP versions, so instead of replacing the original I've augmented it, just to make sure.